separate maps and jukebox to own plugins
