+++ core/plugins/maps/mapsWindow.ts
@@ -21,7 +21,7 @@ export default class MapsWindow extends ListWindow {
                     Object.assign(map, {
                         Index: i++,
                         Name: escape(map.Name),
-                        Author: map.AuthorNickname || map.Author || "",
+                        Author: escape(map.AuthorNickname || map.Author || ""),
                         AuthorTime: formatTime(map.AuthorTime || map.GoldTime),
                     })
                 );
